feat: :sparkles: adding milipede files module
